move leaf nodes of ocm
