--- pkg/unidling/OWNERS
@@ -1,5 +1,0 @@
-reviewers:
-  - DirectXMan12
-  - soltysh
-approvers:
-  - soltysh
