--- pkg/unidling/doc.go
@@ -1,2 +1,0 @@
-// Package unidling supports unidling scalable resources upon network traffic to associated services
-package unidling
